Added loading in data functionality.
Author(s):	Preston Ward, Anthony Pizzimenti
Date:		June 12, 2018

Edited the ingest to no longer require a congressional district but
instead asks for the GEOID column.  Added a test data folder of the fake
shapefile stuff and data.  Added the functionality to be able to specify a
CVS or shp file and column name to then pull that into a list for adding
to your nodes.