Bug fixes: Off by 1 on 0 VTD/CD Assignments
Fixed the issue where we were sometimes off by 1 if the VTD at position
zero was referenced.  There was also an issue where the nodes being
assigned data had changed their order to how the neighbors were
constructed and not the order of the data table.  Data now assigns to the
proper node.