Cleanup neighbor selection in ingest.py
The results of this should be mostly identical to the previous version,
including the (probably wrong) perimeter calculations. The neighbors are
now returned as a dictionary, which is only superficially different for
this code.

Improve ingest.py documentation

Reduce complexity of ingest.py

This also slightly modifies `construct_graph()` to account for using a
neighbor dictionary.

Improve return docstring on `ingest`