Compute shared perimeters, streamline ingestion
Previously, ingest was unnecessarily unpacking the pysal weights into an adjacency matrix, but it is conveniently already provided in a dictionary. So I streamlined it.

More importantly, the previous version was not actually computing the shared perimeters between neighbors, but just listing the full perimeter of each shape.

I’ve made it so the output is identical to before (with the corrected shared perimeter values), but I think that the sorting is unnecessary and that the plain neighbors and shared_perims dicts could be handed off to the networkx generation function in their original form, and that integer IDs could be assigned there. Probably doesn’t matter much, but you could potentially have an indexing headache somewhere.