Logs statistics while moving through the chain.
Author(s):      Anthony Pizzimenti, Mary Barker
Date:           June 14, 2018

Wrapper class for instances of MarkovChain. As is specified in
__main__.py, main() now exposes the MarkovChain instance. Then, the
Logger can consume it, run it, and do any specified logging. Currently,
the statistics of each state the chain moves to are binned and plotted
after the chain is complete.

The Logger class can (and definitely should) be augmented, as Logger
itself runs the chain.