Add optional max_attempts argument to bipartition tree funcs (#389)
* Add optional max_attempts argument to bipartition tree funcs

This patch adds a max_attempts argument to the bipartition_tree funcs in
`gerrychain.tree`. If max_attempts is None, it will default to the
previous behavior (i.e. it will hang if it gets stuck). If max_attempts
is set, it will throw a RuntimeError when `max_attempts` is reached.

This patch is based off code from Robi Huq (via Jeanne Clelland), but
is rewritten. Any mistakes are likely my own.

* Bump up default max_attempts to 10k, just in case

Co-authored-by: Robi Huq <robiohuq@gmail.com>