Since returning from vacation, I have been focused on getting all of the tests to pass (as well as writing tests for new code added to graph.py).  There is a little bit of general cleanup as well.
A good bit of the changes were just to deal with NX node_ids in tests
that required translation to or from RX node_ids or to / from what
I refer to as "original" node_ids which are the NX node_ids given
to the original NX graph.

I had to essentially duplicate the tests in test_tree.py to test
RX-based Graph objects in addition to NX-based Graph objects.

I made an executive decision (which you can un-decide) concerning
Election objects.  The original code allowed creation of an
Election object with an explicit map of {node_id: vote_tally_integer}.
This was then used to create the initial DataTally objects, which
was problematic because the node_ids were NX node_ids which were
not appropriate for tallying data in the RX-based graph that
was in the partition.  So, I just added code to the Election
class prohibiting the use of node_ids - instead you can only
use attribute names.  I checked and there was no other use
of explicit node_id to vote count data in any of the tests or
in any of the code examples you sent me (note that I did not
look at the User Guide or any other example notebook code).
I then changed the one test case to use attribute names...